system-plugin: move root plugin files into plugin/ directory
