+++ src/plugin/common.h
