+++ src/startup/load.h
@@ -13,7 +13,7 @@
 #ifndef SYSTEM_PLUGIN_STARTUP_LOAD_H
 #define SYSTEM_PLUGIN_STARTUP_LOAD_H
 
-#include "context.h"
+#include "plugin/context.h"
 
 int system_startup_load_data(system_ctx_t *ctx, sr_session_ctx_t *session);
