+++ src/startup/store.h
@@ -13,7 +13,7 @@
 #ifndef SYSTEM_PLUGIN_STARTUP_STORE_H
 #define SYSTEM_PLUGIN_STARTUP_STORE_H
 
-#include "context.h"
+#include "plugin/context.h"
 
 int system_startup_store_data(system_ctx_t *ctx, sr_session_ctx_t *session);
