+++ src/system/api/load.h
@@ -13,8 +13,8 @@
 #ifndef SYSTEM_PLUGIN_API_LOAD_H
 #define SYSTEM_PLUGIN_API_LOAD_H
 
-#include "common.h"
-#include "context.h"
+#include "plugin/common.h"
+#include "plugin/context.h"
 
 int system_load_hostname(system_ctx_t *ctx, char buffer[SYSTEM_HOSTNAME_LENGTH_MAX]);
 int system_load_contact(system_ctx_t *ctx, char buffer[256]);
