+++ src/system/api/ntp/change.h
@@ -13,8 +13,8 @@
 #ifndef SYSTEM_PLUGIN_API_NTP_CHANGE_H
 #define SYSTEM_PLUGIN_API_NTP_CHANGE_H
 
-#include "types.h"
-#include "context.h"
+#include "plugin/types.h"
+#include "plugin/context.h"
 
 #include <srpc.h>
