+++ src/system/api/ntp/store.h
@@ -13,8 +13,8 @@
 #ifndef SYSTEM_PLUGIN_API_NTP_STORE_H
 #define SYSTEM_PLUGIN_API_NTP_STORE_H
 
-#include "types.h"
-#include "context.h"
+#include "plugin/types.h"
+#include "plugin/context.h"
 
 int system_ntp_store_server(system_ctx_t *ctx, system_ntp_server_element_t *head);
