+++ src/system/data/authentication/authorized_key.h
@@ -13,7 +13,7 @@
 #ifndef SYSTEM_PLUGIN_DATA_AUTHENTICATION_AUTHORIZED_KEY_H
 #define SYSTEM_PLUGIN_DATA_AUTHENTICATION_AUTHORIZED_KEY_H
 
-#include "types.h"
+#include "plugin/types.h"
 
 void system_authorized_key_init(system_authorized_key_t *key);
 int system_authorized_key_set_name(system_authorized_key_t *key, const char *name);
