+++ src/system/data/dns_resolver/search.h
@@ -13,7 +13,7 @@
 #ifndef SYSTEM_PLUGIN_DATA_DNS_RESOLVER_SEARCH_H
 #define SYSTEM_PLUGIN_DATA_DNS_RESOLVER_SEARCH_H
 
-#include "types.h"
+#include "plugin/types.h"
 
 void system_dns_search_init(system_dns_search_t *search);
 int system_dns_search_set_domain(system_dns_search_t *search, const char *domain);
