+++ src/system/data/dns_resolver/server/list.h
@@ -13,7 +13,7 @@
 #ifndef SYSTEM_PLUGIN_DATA_DNS_RESOLVER_SERVER_LIST_H
 #define SYSTEM_PLUGIN_DATA_DNS_RESOLVER_SERVER_LIST_H
 
-#include "types.h"
+#include "plugin/types.h"
 
 void system_dns_server_list_init(system_dns_server_element_t **head);
 int system_dns_server_list_add(system_dns_server_element_t **head, system_dns_server_t server);
